Density Function Code Refactor, Implement Depth Value
