+++ src/main/java/com/aetherteam/aetherii/data/resources/builders/AetherIIDensityFunctionBuilders.java
@@ -1,0 +1,53 @@
+package com.aetherteam.aetherii.data.resources.builders;
+
+import com.aetherteam.aetherii.AetherII;
+import com.aetherteam.aetherii.data.resources.registries.AetherIIDensityFunctions;
+import net.minecraft.core.HolderGetter;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.levelgen.DensityFunction;
+import net.minecraft.world.level.levelgen.DensityFunctions;
+
+public class AetherIIDensityFunctionBuilders {
+    public static final ResourceKey<DensityFunction> TEMPERATURE = createKey("highlands/temperature");
+    public static final ResourceKey<DensityFunction> EROSION = createKey("highlands/erosion");
+    public static final ResourceKey<DensityFunction> ELEVATION = createKey("highlands/elevation");
+    public static final ResourceKey<DensityFunction> FACTOR = createKey("highlands/factor"); //TODO: Add to Datagen
+    public static final ResourceKey<DensityFunction> BOTTOM_SLIDE = createKey("highlands/bottom_slide"); //TODO: Add to Datagen
+    public static final ResourceKey<DensityFunction> TOP_SLIDE = createKey("highlands/top_slide"); //TODO: Add to Datagen
+    public static final ResourceKey<DensityFunction> SLOPER = createKey("highlands/sloper"); //TODO: Add to Datagen
+    public static final ResourceKey<DensityFunction> BASE_3D_NOISE = createKey("highlands/base_3d_noise");
+    public static final ResourceKey<DensityFunction> AMPLIFICATION = createKey("highlands/amplification"); //TODO: Add to Datagen
+    public static final ResourceKey<DensityFunction> TERRAIN_SHAPER = createKey("highlands/terrain_shaper"); //TODO: Add to Datagen
+
+    private static ResourceKey<DensityFunction> createKey(String name) {
+        return ResourceKey.create(Registries.DENSITY_FUNCTION, new ResourceLocation(AetherII.MODID, name));
+    }
+
+    public static final ResourceKey<DensityFunction> SHIFT_X = createVanillaKey("shift_x");
+    public static final ResourceKey<DensityFunction> SHIFT_Z = createVanillaKey("shift_z");
+    public static final ResourceKey<DensityFunction> Y = createVanillaKey("y");
+
+    private static ResourceKey<DensityFunction> createVanillaKey(String name) {
+        return ResourceKey.create(Registries.DENSITY_FUNCTION, new ResourceLocation(name));
+    }
+
+    public static DensityFunction factorize(HolderGetter<DensityFunction> function, double value) {
+        DensityFunction density = getFunction(function, AetherIIDensityFunctions.FACTOR);
+        density = DensityFunctions.mul(density, DensityFunctions.constant(value));
+        density = DensityFunctions.mul(density, getFunction(function, AetherIIDensityFunctions.TERRAIN_SHAPER));
+        return density;
+    }
+
+    public static DensityFunction makeTerrainShaper(HolderGetter<DensityFunction> function) {
+        DensityFunction density = getFunction(function, AMPLIFICATION);
+        density = DensityFunctions.add(density, DensityFunctions.yClampedGradient(96, 128, 0.75, 0.35));
+        density = DensityFunctions.mul(density, getFunction(function, SLOPER));
+        return density.clamp(0, 1);
+    }
+
+    public static DensityFunction getFunction(HolderGetter<DensityFunction> densityFunctions, ResourceKey<DensityFunction> key) {
+        return new DensityFunctions.HolderHolder(densityFunctions.getOrThrow(key));
+    }
+}
